Add enums.py
